requirements.txt
streamlit==1.8.0
psycopg2-binary==2.9.3
pandas==1.4.1
beautifulsoup4==4.10.0
requests==2.27.1
python-dotenv==0.19.2
